+++ groupeB/trunk/src/main/java/fr/alma/server/core/IEmplacement.java
@@ -3,8 +3,8 @@ package fr.alma.server.core;
 import java.util.List;
 
 public interface IEmplacement {
-	short getCol();
-	short getRow();
+	int getCol();
+	int getRow();
 	public boolean isIn(List<IEmplacement> emplacements);
 	boolean equals(IEmplacement e1);
